+++ CHANGES.rst
@@ -1,3 +1,21 @@
+ChangeLog
+=========
+
+0.0.2 (2017-04-15)
+------------------
+
+Minor Enhancements + Doc Updates
+
+Changes
+*******
+- Documentation updated to emphasize the need for supplying a client token
+  and secret when using the library (in addition to the username and password
+  credentials for interacting on behalf of a user)
+
+Features
+********
+- Access tokens are automatically refreshed when they expire
+
 0.0.1 (2017-04-14)
 ------------------